Implemented multi layered type names
* so it loads a default device if no other device is found
* that adds the functionality, that a model can be later be specified better and still load correctly afterwards
* removed unused usings
* removed play store entry from about page
* removed unused models, that where needed in the past
