+++ ios/Flutter/flutter_export_environment.sh
@@ -5,8 +5,8 @@ export "FLUTTER_APPLICATION_PATH=C:\Users\susch\source\flutter\smarthome"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib\main.dart"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=1.1.2"
-export "FLUTTER_BUILD_NUMBER=1"
+export "FLUTTER_BUILD_NAME=1.1.3"
+export "FLUTTER_BUILD_NUMBER=0"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=false"
 export "TREE_SHAKE_ICONS=false"
