+++ lib/devices/device_exporter.dart
@@ -7,10 +7,8 @@ export 'xiaomi/temp_sensor.dart';
 export 'base_model.dart';
 export 'xiaomi/temp_sensor_model.dart';
 export 'floaltpanel/floalt_panel.dart';
-export 'floaltpanel/floalt_panel_model.dart';
 export 'zigbee/zigbee_model.dart';
 export 'osramb40rw/osram_b40_rw.dart';
-export 'osramb40rw/osram_b40_rw_model.dart';
 export 'osramplug/osram_plug.dart';
 export 'osramplug/osram_plug_model.dart';
 export 'tradfriledbulb/tradfri_led_bulb.dart';
@@ -19,3 +17,5 @@ export 'tradfricontroloutlet/tradfri_control_outlet.dart';
 export 'tradfricontroloutlet/tradfri_control_outlet_model.dart';
 export 'tradfrimotionsensor/tradfri_motion_sensor.dart';
 export 'tradfrimotionsensor/tradfri_motion_sensor_model.dart';
+export 'zigbee/zigbeelamp/zigbee_lamp.dart';
+export 'zigbee/zigbeelamp/zigbee_lamp_model.dart';
