+++ lib/devices/floaltpanel/floalt_panel.dart
@@ -1,194 +1,20 @@
 // ignore_for_file: unnecessary_null_comparison
 
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 // import 'package:signalr_client/signalr_client.dart';
 import 'package:signalr_core/signalr_core.dart';
-import 'package:smarthome/controls/gradient_rounded_rect_slider_track_shape.dart';
-import 'package:smarthome/devices/device.dart';
 import 'package:smarthome/devices/device_manager.dart';
-import 'package:smarthome/helper/theme_manager.dart';
-import 'package:smarthome/models/message.dart' as sm;
+import 'package:smarthome/devices/zigbee/zigbeelamp/zigbee_lamp.dart';
+import 'package:smarthome/devices/zigbee/zigbeelamp/zigbee_lamp_model.dart';
 
 import '../device_manager.dart';
-import 'floalt_panel_model.dart';
 
-class FloaltPanel extends Device<FloaltPanelModel> {
-  FloaltPanel(int? id, String typeName, FloaltPanelModel model, HubConnection connection, IconData icon)
+class FloaltPanel extends ZigbeeLamp {
+  FloaltPanel(int? id, String typeName, ZigbeeLampModel model, HubConnection connection, IconData icon)
       : super(id, typeName, model, connection, icon);
-
-  @override
-  void navigateToDevice(BuildContext context) {
-    Navigator.push(context, MaterialPageRoute(builder: (BuildContext context) => FloaltPanelScreen(this)));
-  }
-
-  @override
-  Widget dashboardCardBody() {
-    return Wrap(
-      alignment: WrapAlignment.center,
-      runAlignment: WrapAlignment.spaceEvenly,
-      children: [
-        MaterialButton(
-          child: Text(
-            "An",
-            style: baseModel.state ? TextStyle(fontWeight: FontWeight.bold, fontSize: 20) : TextStyle(),
-          ),
-          onPressed: () => sendToServer(sm.MessageType.Update, sm.Command.SingleColor, []),
-        ),
-        MaterialButton(
-          child: Text(
-            "Aus",
-            style: !baseModel.state ? TextStyle(fontWeight: FontWeight.bold, fontSize: 20) : TextStyle(),
-          ),
-          onPressed: () => sendToServer(sm.MessageType.Update, sm.Command.Off, []),
-        ),
-      ],
-    );
-  }
 
   @override
   DeviceTypes getDeviceType() {
     return DeviceTypes.FloaltPanel;
-  }
-}
-
-class FloaltPanelScreen extends DeviceScreen {
-  final FloaltPanel floaltPanel;
-  FloaltPanelScreen(this.floaltPanel);
-
-  @override
-  State<StatefulWidget> createState() => _FloaltPanelScreenState();
-}
-
-class _FloaltPanelScreenState extends State<FloaltPanelScreen> {
-  DateTime dateTime = DateTime.now();
-  late StreamSubscription sub;
-
-  void sliderChange(Function f, int dateTimeMilliseconds, [double? val]) {
-    if (DateTime.now().isAfter(dateTime.add(new Duration(milliseconds: dateTimeMilliseconds)))) {
-      Function.apply(f, val == null ? [] : [val]);
-      dateTime = DateTime.now();
-    }
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    sub = this.widget.floaltPanel.listenOnUpdateFromServer((p0) {
-      setState(() {});
-    });
-  }
-
-  @override
-  void deactivate() {
-    super.deactivate();
-    sub.cancel();
-  }
-
-  void changeDelay(double? delay) {
-    this.widget.floaltPanel.sendToServer(sm.MessageType.Options, sm.Command.Delay, [delay.toString()]);
-  }
-
-  void changeBrightness(double brightness) {
-    this.widget.floaltPanel.sendToServer(sm.MessageType.Update, sm.Command.Brightness, [brightness.round().toString()]);
-  }
-
-  void changeColorTemp(double colorTemp) {
-    this.widget.floaltPanel.sendToServer(sm.MessageType.Update, sm.Command.Temp, [colorTemp.round().toString()]);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return new Scaffold(
-      appBar: AppBar(
-        title: new Text(this.widget.floaltPanel.baseModel.friendlyName),
-      ),
-      body: Container(
-        decoration: ThemeManager.getBackgroundDecoration(context),
-        child: buildBody(this.widget.floaltPanel.baseModel),
-      ),
-      floatingActionButton: FloatingActionButton(
-        child: const Icon(Icons.power_settings_new),
-        onPressed: () => this.widget.floaltPanel.sendToServer(sm.MessageType.Update, sm.Command.Off, []),
-      ),
-    );
-  }
-
-  Widget buildBody(FloaltPanelModel model) {
-    return ListView(
-      children: <Widget>[
-        ListTile(
-          title: Text("Angeschaltet: " + (model.state ? "Ja" : "Nein")),
-        ),
-        ListTile(
-          title: Text("Verfügbar: " + (model.available ? "Ja" : "Nein")),
-        ),
-        ListTile(
-          title: Text("Verbindungsqualität: " + (model.linkQuality.toString())),
-        ),
-        ListTile(
-          title: Text("Helligkeit " + model.brightness.toStringAsFixed(0)),
-          subtitle: GestureDetector(
-            child: SliderTheme(
-              child: Slider(
-                value: model.brightness.toDouble(),
-                onChanged: (d) {
-                  setState(() => model.brightness = d.round());
-                  sliderChange(changeBrightness, 500, d);
-                },
-                min: 0.0,
-                max: 100.0,
-                divisions: 100,
-                label: '${model.brightness}',
-              ),
-              data: SliderTheme.of(context).copyWith(
-                  trackShape: GradientRoundedRectSliderTrackShape(
-                      LinearGradient(colors: [Colors.grey.shade800, Colors.white]))),
-            ),
-            onTapCancel: () => changeBrightness(model.brightness.toDouble()),
-          ),
-        ),
-        ListTile(
-          title: Text("Farbtemparatur " + (model.colorTemp - 204).toStringAsFixed(0)),
-          subtitle: GestureDetector(
-            child: SliderTheme(
-              child: Slider(
-                value: ((model.colorTemp - 204).clamp(0, 204)).toDouble(),
-                onChanged: (d) {
-                  setState(() => model.colorTemp = d.round() + 204);
-                  sliderChange(changeColorTemp, 500, d + 204.0);
-                },
-                min: 0.0,
-                max: 204.0,
-                divisions: 204,
-                label: '${model.colorTemp - 204}',
-              ),
-              data: SliderTheme.of(context).copyWith(
-                  trackShape: GradientRoundedRectSliderTrackShape(
-                      LinearGradient(colors: [Color.fromARGB(255, 255, 209, 163), Color.fromARGB(255, 255, 147, 44)]))),
-            ),
-            onTapCancel: () => changeColorTemp(model.colorTemp.toDouble()),
-          ),
-        ),
-        ListTile(
-          title: Text("Übergangszeit " + model.transitionTime.toStringAsFixed(1) + " Sekunden"),
-          subtitle: GestureDetector(
-            child: Slider(
-              value: model.transitionTime,
-              onChanged: (d) {
-                setState(() => model.transitionTime = d);
-                sliderChange(changeDelay, 500, d);
-              },
-              min: 0.0,
-              max: 10.0,
-              divisions: 100,
-              label: '${model.transitionTime}',
-            ),
-            onTapCancel: () => changeDelay(model.transitionTime),
-          ),
-        ),
-      ],
-    );
   }
 }
